Move ioloop based client to new module, improve
